# 7agz.api/bin/Release/net6.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "7agz.api, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Reservation/FindStadium": {
      "get": {
        "tags": [
          "Reservation"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "latitude",
            "type": "number",
            "format": "double"
          },
          {
            "in": "query",
            "name": "longitude",
            "type": "number",
            "format": "double"
          },
          {
            "in": "query",
            "name": "distance",
            "type": "number",
            "format": "double",
            "default": 1000
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Reservation/ReserveStadium": {
      "post": {
        "tags": [
          "Reservation"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/ReservationDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Reservation/ReserveStadiumOwner": {
      "post": {
        "tags": [
          "Reservation"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/ReservationOwnerDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Reservation/FindTeam": {
      "get": {
        "tags": [
          "Reservation"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "latitude",
            "type": "number",
            "format": "double"
          },
          {
            "in": "query",
            "name": "longitude",
            "type": "number",
            "format": "double"
          },
          {
            "in": "query",
            "name": "distance",
            "type": "number",
            "format": "double",
            "default": 1000
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Reservation/JoinTeam": {
      "post": {
        "tags": [
          "Reservation"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "playerId",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "teamId",
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Reservation/WithdrawTeam": {
      "delete": {
        "tags": [
          "Reservation"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "playerId",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "teamId",
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Reservation/CancelReservation": {
      "delete": {
        "tags": [
          "Reservation"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "hourId",
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Stadium/AddStadium": {
      "post": {
        "tags": [
          "Stadium"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/StadiumDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Stadium/GetSatdiums": {
      "get": {
        "tags": [
          "Stadium"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Stadium/{id}": {
      "patch": {
        "tags": [
          "Stadium"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Operation"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Stadium"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Stadium/id": {
      "put": {
        "tags": [
          "Stadium"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/StadiumDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Stadium/GetReservedHours": {
      "get": {
        "tags": [
          "Stadium"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "definitions": {
...
    "Operation": {
      "type": "object",
      "properties": {
        "operationType": {
          "$ref": "#/definitions/OperationType"
        },
        "path": {
          "type": "string"
        },
        "op": {
          "type": "string"
        },
        "from": {
          "type": "string"
        },
        "value": { }
      }
    },
    "OperationType": {
      "format": "int32",
      "enum": [
        0,
        1,
        2,
        3,
        4,
        5,
        6
      ],
      "type": "integer"
    },
...
    "ReservationDTO": {
      "type": "object",
      "properties": {
        "stadiumId": {
          "format": "int32",
          "type": "integer"
        },
        "reservedHour": {
          "format": "date-time",
          "type": "string"
        },
        "reserverId": {
          "format": "int32",
          "type": "integer"
        },
        "reserverType": {
          "type": "boolean"
        },
        "teamMembers1": {
          "type": "array",
          "items": {
            "format": "int32",
            "type": "integer"
          }
        },
        "teamMembers2": {
          "type": "array",
          "items": {
            "format": "int32",
            "type": "integer"
          }
        }
      }
    },
    "ReservationOwnerDTO": {
      "type": "object",
      "properties": {
        "stadiumId": {
          "format": "int32",
          "type": "integer"
        },
        "reservedHour": {
          "format": "date-time",
          "type": "string"
        },
        "reserverId": {
          "format": "int32",
          "type": "integer"
        },
        "reserverType": {
          "type": "boolean"
        }
      }
    },
    "StadiumDTO": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "start": {
          "type": "string"
        },
        "end": {
          "type": "string"
        },
        "hourPrice": {
          "format": "int32",
          "type": "integer"
        },
        "longitude": {
          "format": "double",
          "type": "number"
        },
        "latitude": {
          "format": "double",
          "type": "number"
        },
        "capacity": {
          "format": "int32",
          "type": "integer"
        },
        "vestiary": {
          "type": "boolean"
        },
        "cafeteria": {
          "type": "boolean"
        },
        "parkingArea": {
          "type": "boolean"
        },
        "stadiumOwnerId": {
          "format": "int32",
          "type": "integer"
        }
      }
    }
  }
}